requirements.txt
Flask
pandas
scikit-learn
matplotlib
requests
python-dotenv